Update response status code on successful deletion
Updated the response status code to return HttpStatus.NO_CONTENT upon successful deletion in book, borrowing, and author API controllers. This is to align with standard practices since "No Content" indicates that the server has successfully fulfilled the request and there is no additional content to send in the response payload body.